+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Page is now a state instead of being defined by `from` and `too`
 
+### Added
+
+- GA documentation
+
+## [0.4.2] - 2020-01-09
+
+### Added
+
+- Query and operator info on redirect return.
+
 ## [0.4.0] - 2019-12-27
 
 ### Added